+++ chaos_theory/docs/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Add `should_log()`, remove `Source::should_log`, make `vdbg!`, `vprintln!`, and `veprintln!` usable everywhere.
+
 ## 0.3.5 (2026-03-25)
 
 - Fix broken `chaos_theory_derive` dependency version.